+++ src/Core/Domain/Product/Attachment/Command/AssociateProductAttachmentsCommand.php
@@ -24,34 +24,30 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Image\Command;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Command;
 
-use PrestaShop\PrestaShop\Core\Domain\Product\Image\DTO\ImageCollection;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 
 /**
- * Updates product images.
+ * Associates products with attachments.
  */
-class UpdateImagesCommand
+class AssociateProductAttachmentsCommand
 {
     /**
      * @var ProductId
      */
     private $productId;
 
-    /**
-     * @var ImageCollection
-     */
-    private $images;
+    private $attachmentIds;
 
     /**
      * @param int $productId
-     * @param ImageCollection $images
+     * @param int[] $attachmentIds
      */
-    public function __construct(int $productId, ImageCollection $images)
+    public function __construct(int $productId, array $attachmentIds)
     {
         $this->productId = new ProductId($productId);
-        $this->images = $images;
+        $this->setAttachmentIds($attachmentIds);
     }
 
     /**
@@ -63,10 +59,31 @@ class UpdateImagesCommand
     }
 
     /**
-     * @return ImageCollection
+     * @return mixed
      */
-    public function getImages(): ImageCollection
+    public function getAttachmentIds()
     {
-        return $this->images;
+        return $this->attachmentIds;
+    }
+
+    private function setAttachmentIds(array $attachmentIds): void
+    {
+        foreach ($attachmentIds as $attachmentId) {
+            //todo: change me when AttachmentId VO is available
+            $this->attachmentIds[] = new class($attachmentId) {
+
+                private $attachmentId;
+
+                public function __construct(int $attachmentId)
+                {
+                    $this->attachmentId;
+                }
+
+                public function getValue()
+                {
+                    return $this->attachmentId;
+                }
+            };
+        }
     }
 }
